aspeed_scu: Add bits to claim DRAM as initialised
Teddy Reed <teddy@prosauce.org> discovered the bit we needed to flip to
claim DRAM initialisation was complete. With this bit set we can avoid
emulating interactions that would otherwise be required to support the
training routines.

This was borrowed from Teddy's parallel introduction of a SCU model,
so also added the copyright and attribution.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>