i2c: add aspeed i2c controller
The Aspeed AST2400 integrates a set of 14 I2C/SMBus bus controllers
directly connected to ab APB bus which can be programmed as a master
or slave.

This patch offers a device model for the master mode only, slave mode
is not supported.

On the TODO list, we also have :

 - improve and harden the state machine. This is really a first
   version.
 - bus recovery support (used by the Linux driver).
 - transfer mode state machine bits. this is not strictly necessary as
   it is mostly used for debug. The bus busy bit is deducted from the
   I2C core engine of qemu.
 - support of the pool buffer: 2048 bytes of internal SRAM (not used
   by the Linux driver).

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>