net: add FTGMAC100 support
The FTGMAC100 device is configured and controlled via a set of
registers available on AHB bus. Packets are transmitted and received
using DMA ring buffers.

The model is for the moment complete enough to satisfy U-Boot and the
Linux kernel but there are areas that need some more attention :

 - PHY support, this is a copy paste from the Coldfire model
 - Automatic receive/transmit polling, no timers are used to poll the
   ring buffers descriptors. We just expect the qemu nic backend to
   push packets when they arrive.
 - CRC

And some are still on the TODO list (not used today in the drivers) :

 - Ethernet Address Filtering
 - Multicast
 - High priority queues
 - VLAN
 - NCSI, may be we could add n/w filters
 - etc.

The code is based on the Coldfire Fast Ethernet Controller emulation.

Signed-off-by: Cédric Le Goater <clg@kaod.org>