m25p80: add a m25p80_set_rom_storage() routine
Some SPI controllers, such as the Aspeed AST2400, have a mode in which
accesses to the flash content are no different than doing MMIOs. The
controller generates all the necessary commands to load (or store)
data in memory.

To emulate such a behavior, we need to have access to the underlying
storage of the flash object. The purpose of the routine is to set the
storage of the preinitialized SPI flash object which can be used by
the object itself but also by a SPI controller to handled the MMIOs.

This seems sufficient to support read only accesses for the
moment. For writes, we would need to externalize flash_write8()
routine certainly.

Signed-off-by: Cédric Le Goater <clg@kaod.org>