m25p80: add a get storage routine
To emulate boot the first flash such a behavior, we need to access the
underlying storage of the m25p80 flash model and fill a ROM region
with the flash content. This is what the m25p80_get_storage() routine
lets us do.

This is a rather hacky but until we can boot from a MMIO region, it
seems difficult to do anything else.

Signed-off-by: Cédric Le Goater <clg@kaod.org>