wdt: aspeed: Fix failed reboot due to timer miscalculation
The implementation was incorrectly scaling the reload value by
nanoseconds per second. In reality the span represented by the reload
value is determined by the selected clock, either APB or 1MHz.

Test the selected clock and apply the appropriate scaling to
nanoseconds for the timer infrastructure. Note that the APB rate is
hard-coded and reflects the value used by the timer controller, and
isn't necessarily an accurate representation of the clock configuration
programmed in the SCU.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>