hw/arm/aspeed: boot from the FMC CE0 flash module
Now that MMIO execution is supported, remove the copy of the flash
module contents in a ROM memory region and boot directly from CE0 of
the FMC controller using a memory region alias.

Booting from an alternate module (FMC CE1) should be possible. There
is still some work to support the module swap when the watchdog resets
the system.

Signed-off-by: Cédric Le Goater <clg@kaod.org>