aspeed: add the VUART serial
The Aspeed SoCs integrate a Virtual UART module providing virtual
serial communication between the host CPU and the SoC CPU. To
activate, just add an extra '-serial' option on the command line.

Communication with a PowerNV guest console can be done with a chardev
device. Start the BMC guest with an extra :

    -chardev socket,id=vuart,host=localhost,port=9003,ipv4,server,nowait \
    -serial chardev:vuart

and the PowerNV guest with :

    -chardev socket,id=vuart,host=localhost,port=9003,reconnect=10 \
    -serial chardev:vuart

The following registers, which can be accessed by the BMC CPU through
APB bus, are not addressed and would need a specific model :

    VUART20 (Slave): General Control Register A
    VUART24 (Slave): General Control Register B
    VUART28 (Slave): VUART Address Register L
    VUART2C (Slave): VUART Address Register H

Signed-off-by: Cédric Le Goater <clg@kaod.org>