aspeed: handle some memory transaction failures
Some firmwares scan the address space with loads and stores to guess
how much RAM the SoC has. But now that we don't ignore anymore memory
transaction failures, we need a way to let the CPU do such memory
accesses.

We use a dummy memory region for this purpose. Its size is calculated
from the configured RAM size and the maximum RAM size of the SoC. It
is mmapped after the RAM and only returns zeros.

Signed-off-by: Cédric Le Goater <clg@kaod.org>