aspeed: add support for the AST2600 eval board
Initial definitions for a simple machine using an AST2600 SoC (Cortex
CPU).

Differences with the AST2400 and the AST2500 SoCs are handled using
the ASPEED_IS_AST2600() macro. This is not optimal but it is not too
invasive either. We could modify the model to add custom instance_init
and realize handlers in the future.

Signed-off-by: Cédric Le Goater <clg@kaod.org>