hw/ssi: aspeed: Deselect chip to reset state in command mode
This is probably a bug elsewhere, maybe in the kernel driver. However,
this workaround/fix appears to resolve the squashfs/UBI corruption we've
been seeing for quite some time in QEMU with e.g. witherspoon-bmc and
associated images.

The problem is the chip is left in e.g. read mode and the deselect is
necessary to reset it back to idle for the current command to issue
correctly. Failure to deselect leads to command-mode access of data at
unexpected addresses (the state mismatch means the command/address are
potentially ignored) and the result presents as data corruption.

Something I don't yet understand is the symptom (squashfs corruption)
wasn't overly deterministic. I would have expected more widespread
symptoms given the change. Having said that, UBI hammers the flash with
its wear-leveling (somewhat ironically). Its approach appears to be
somewhat complex, so maybe it's not surprising that the results are so
variable.

Cc: Andrew Geissler <geissonator@gmail.com>
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Cédric Le Goater <clg@kaod.org>